[Playlist] Done function delete

+++ Wisky/Controllers/MediaSrcController.cs
@@ -92,7 +92,32 @@ namespace DSS.Controllers
             }
             return mediaSrcUseVMs;
         }
-
+        //TOANTXSE
+        //Get media Src List by Brand ID
+        public static List<Models.MediaSrcUseVM> GetMediaSrcListByBrandIdAndStatus()
+        {
+            IMediaSrcService mediaSrcService = DependencyUtils.Resolve<IMediaSrcService>();
+            var mediaSrcUseVMs = new List<Models.MediaSrcUseVM>();
+            var user = Helper.GetCurrentUser();
+            var mediaSrcList = mediaSrcService.GetMediaSrcByBrand(user.BrandID);
+            foreach (var item in mediaSrcList)
+            {
+                if(item.isPublic == true)
+                {
+                    var m = new Models.MediaSrcUseVM
+                    {
+                        Title = item.Title,
+                        Description = item.Description,
+                        URL = item.URL,
+                        isPublic = (bool)item.isPublic,
+                        MediaSrcId = item.MediaSrcID,
+                        TypeId = item.TypeID,
+                    };
+                    mediaSrcUseVMs.Add(m);
+                }
+            }
+            return mediaSrcUseVMs;
+        }
         // Media/Add resource
         //[HttpPost]
         //public async System.Threading.Tasks.Task<ActionResult> Add(Models.MediaSrcVM model, HttpPostedFileBase file)
@@ -233,12 +258,68 @@ namespace DSS.Controllers
             return 0;
         }
 
+        //TOANTXSE
+        // POST: Media/CheckMediaIdIsUsed  
+        [HttpPost]
+        public JsonResult CheckMediaIdIsUsed(int id)
+        {
+            try
+            {
+                //Get device by screen Id
+                IPlaylistItemService playlistItemService = DependencyUtils.Resolve<IPlaylistItemService>();
+                IPlaylistService playlistService = DependencyUtils.Resolve<IPlaylistService>();
+                var playlistItem = playlistItemService.Get().ToList();
+                var playlist = playlistService.Get().ToList();
+                var playlistItemVMs = new List<Models.PlaylistItemVM>();
+                var playlistVMs = new List<Models.PlaylistDetailVM>();
+                //check mediaSrcId have in playlistItem
+                foreach (var item in playlistItem)
+                {
+                    if (item.MediaSrcID == id)
+                    {
+                        var b = new Models.PlaylistItemVM
+                        {
+                            playlistId = item.PlaylistID,
 
+                        };
+                        playlistItemVMs.Add(b);
+                    }
+                }
+                // if playlistItemVMs != null, get Playlist Title by playlistID
+                if (playlistItemVMs.Count != 0)
+                {
+                    foreach (var item in playlistItemVMs)
+                    {
+                        foreach (var itemPlaylist in playlist)
+                        {
+                            if (item.playlistId == itemPlaylist.PlaylistID)
+                            {
+                                var b = new Models.PlaylistDetailVM
+                                {
+                                    Title = itemPlaylist.Title,
+                                };
+                                playlistVMs.Add(b);
+                            }
+                        }
+                    }
+                }
+                return Json(new
+                {
+                    isUsing = playlistItemVMs.Count != 0,
+                    playlistVMlist = playlistVMs,
+                }, JsonRequestBehavior.AllowGet);
+            }
+            catch (Exception ex)
+            {
+                throw ex;
+            }
+        }
         // GET: Media/Delete/:id
         public ActionResult Delete(int id)
         {
+            var user = Helper.GetCurrentUser();
             var mediaSrc = this.mediaSrcService.Get(id);
-            if (mediaSrc != null)
+            if (mediaSrc != null && mediaSrc.BrandID == user.BrandID)
             {
                 this.mediaSrcService.Delete(mediaSrc);
             }
